# [MKL-DNN] Fix to int8 performance regression (#18698)