Changes here are restoring performance of int8 convolution caused by muli-threading mechanisms introduction (issue #18698). 

Internal testing proved that there is performance improvement for int8 convolutional models