@luotao1 2019-07-25
I test it on 6271 cpu with Mobilenet-V1/V2 only.
```
./test_analyzer_int8_image_classification --infer_data=int8_full_val.bin --warmup --batch_size=1 --paddle_num_threads=1 --infer_model=int8v2/mobilenetv1/model/
```
### Mobilenet-v1
- before:
```
I0725 13:02:23.514694 341334 tester_helper.h:462] --- Performance summary ---
I0725 13:02:23.514722 341334 tester_helper.h:463] FP32: avg fps: 80.6569, avg latency: 12.3982 ms
I0725 13:02:23.514744 341334 tester_helper.h:466] INT8: avg fps: 197.8102, avg latency: 5.0554 ms
I0725 13:02:23.524044 341334 tester_helper.h:447] --- Accuracy summary ---
I0725 13:02:23.524067 341334 tester_helper.h:448] Accepted top1 accuracy drop threshold: 0.01. (condition: (FP32_top1_acc - INT8_top1_acc) <= threshold)
I0725 13:02:23.524085 341334 tester_helper.h:451] FP32: avg top1 accuracy: 0.7078
I0725 13:02:23.524091 341334 tester_helper.h:453] INT8: avg top1 accuracy: 0.7041
```
- after:
```
I0725 14:03:20.350790 92589 tester_helper.h:462] --- Performance summary ---
I0725 14:03:20.350821 92589 tester_helper.h:463] FP32: avg fps: 77.4833, avg latency: 12.9060 ms
I0725 14:03:20.350841 92589 tester_helper.h:466] INT8: avg fps: 219.8935, avg latency: 4.5477 ms
I0725 14:03:20.357447 92589 tester_helper.h:447] --- Accuracy summary ---
I0725 14:03:20.357475 92589 tester_helper.h:448] Accepted top1 accuracy drop threshold: 0.01. (condition: (FP32_top1_acc - INT8_top1_acc) <= threshold)
I0725 14:03:20.357496 92589 tester_helper.h:451] FP32: avg top1 accuracy: 0.7078
I0725 14:03:20.357501 92589 tester_helper.h:453] INT8: avg top1 accuracy: 0.7041
```
### Mobilenet-v2