@luotao1 2019-07-25
It seems the `before` performance is better than https://github.com/PaddlePaddle/Paddle/issues/18698#issue-470205380. Is this normal @bingyanghuang @jczaja