@bingyanghuang 2019-07-25
> It seems the `before` performance is better than [#18698 (comment)](https://github.com/PaddlePaddle/Paddle/issues/18698#issue-470205380). Is this normal @bingyanghuang @jczaja

So it seems that the performance regression problem has **not** been solved. @jczaja