@jczaja 2019-07-26
@luotao1 I have made some measures using mobilnet v1 on SKX (8180), I will share them , perhaps this will help in understanding situation we are dealing with. 

1. changes before regression (one before commit 84bb45c ). 
batch latency: 6.12171ms, number of samples: 50000, sample latency: 6.12171ms, **fps: 163.353**, data type: int8_t
2. regression commit ( commit 84bb45c ):
 batch latency: 8.22298ms, number of samples: 50000, sample latency: 8.22298ms, **fps: 121.61**, data type: int8_t 
3. regression commit + this PR (ported changes to regression commit):
batch latency: 6.26656ms, number of samples: 50000, sample latency: 6.26656ms, **fps: 159.577**, data type: int8_t

4. Recent develop ( 25th of July):
 INT8: **avg fps: 150.7610**, avg latency: 6.6330 ms    

5. Recent develop + PR :
INT8: **avg fps: 159.3892**, avg latency: 6.2740 ms 

Comments:
a) From point 1 and 2 we can see that regression was from 163 fps -> 121 fps . Then in 3
we see that improvement is 121 fps -> 159 fps . Why not 163 ? Answer is that generation of string from Thread Id via stringstream caused 4 FPS penalty. To prove that I replaced stringstream with fixed value string (I have run analyzer in single threaded, so it was functionally fine). And performance was again 163 fps. To combat this problem I have updated this PR replacing stringstream usage with to_string(std::hash<std::thread::id>)() to get string out of TID.
I think there is small improvement 159 -> ~160 . 

b) You asked why recent before is much better that the one reported. If you look into result from point 2 (regression commit)  and recent develop (regression present, but recent codebase) we can see that there is an improvement 121 fps -> 150 fps. Why is that? I haven't traced it down, but possible reason is that we updated to mkl-dnn 0.20 , and mkl-dnn team reduced time of creation of primitives. So then performance gain is not that big with caching when overhead from creation is smaller. This is hypothesis (I looked at some of their commits, presentations so I know they worked on that) to be confirmed. 

c) From 4 and 5 we can see that there is improvement 150 fps -> 159 fps. Those results are taken on SKX , on next generation (CLX) that you are using difference should be a bit bigger.

Please continue reviewing those changes.