@jczaja 2019-07-29
@luotao1 Could you please advise on how to pass: "PR_CI_Coverage" and "PR_CI_Document_Preview"