@luotao1 2019-07-29
You can ignore these two CI since they are not required now.