@luotao1 2019-07-30
- mobilenetV1: from 219.8935 (https://github.com/PaddlePaddle/Paddle/pull/18758#issuecomment-514914670) ->226.5409
```
I0730 12:18:00.623914 357906 tester_helper.h:463] FP32: avg fps: 78.4434, avg latency: 12.7480 ms
I0730 12:18:00.623934 357906 tester_helper.h:466] INT8: avg fps: 226.5409, avg latency: 4.4142 ms
```
- mobilenetv2: from 191.5456 (https://github.com/PaddlePaddle/Paddle/pull/18758#issuecomment-514914670) -> 201.2669
```
I0730 12:36:33.656062 269110 tester_helper.h:463] FP32: avg fps: 91.1440, avg latency: 10.9717 ms
I0730 12:36:33.656085 269110 tester_helper.h:466] INT8: avg fps: 201.2669, avg latency: 4.9685 ms
```