@luotao1 2019-07-30
LGTM！And the last question 
> Why not 163 ? Answer is that generation of string from Thread Id via stringstream caused 4 FPS penalty. 

Will it be solved after refactoring conv Op?