@jczaja 2019-07-30
@luotao1 Key generation using TID is not related to conv refactoring.  But I would like to try to speed it up , and this effort will be tracked with #18902